+++ src/main/java/com/icare/service/CityService.java
@@ -1,13 +1,16 @@
 package com.icare.service;
 
-import com.icare.dao.entity.CityEntity;
-import com.icare.dao.repository.CityRepository;
+import com.icare.entity.CityEntity;
+import com.icare.repository.CityRepository;
 import com.icare.mapper.CityMapper;
-import com.icare.model.dto.request.CityRequest;
-import com.icare.model.dto.response.CityResponse;
-import com.icare.model.exception.NotFoundException;
+import com.icare.dto.request.CityRequest;
+import com.icare.dto.response.CityResponse;
+import com.icare.exception.DeletionException;
+import com.icare.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,38 +22,33 @@ public class CityService {
     private final CityRepository cityRepository;
 
     public void save(CityRequest request){
-        log.info("ActionLog.save.start for category name is {}", request.getName());
         cityRepository.save(CityMapper.INSTANCE.requestToEntity(request));
-        log.info("ActionLog.save.end for category name is {}", request.getName());
     }
 
-    public List<CityResponse> getAll(){
-        return CityMapper.INSTANCE.entitiesToResponses(cityRepository.findAll());
+    public List<CityResponse> getAllCities(Pageable pageable){
+        Page<CityEntity> entities = cityRepository.findAll(pageable);
+        return CityMapper.INSTANCE.entitiesToResponses(entities);
     }
 
     public CityResponse getById(Long id){
-        CityEntity entity = cityRepository.findById(id).orElseThrow(() -> {
-            log.error("ActionLog.getById.NotFoundException for category id = {}", id);
-            return new NotFoundException("CITY_NOT_FOUND");
-        });
+        CityEntity entity = cityRepository.findById(id).orElseThrow(() ->
+            new NotFoundException("CITY_NOT_FOUND"));
         return CityMapper.INSTANCE.entityToResponse(entity);
     }
 
     public CityResponse editById(Long id, CityRequest request){
-        log.info("ActionLog.editById.start for category id = {}", id);
-        CityEntity entity = cityRepository.findById(id).orElseThrow(() -> {
-            log.error("ActionLog.editById.NotFoundException for category id = {}", id);
-            return new NotFoundException("CITY_NOT_FOUND");
-        });
+        CityEntity entity = cityRepository.findById(id).orElseThrow(() ->
+            new NotFoundException("CITY_NOT_FOUND"));
         CityMapper.INSTANCE.mapRequestToEntity(entity, request);
-        log.info("ActionLog.editById.end for category id = {}", id);
         return CityMapper.INSTANCE.entityToResponse(entity);
     }
 
     public void deleteById(Long id){
-        log.info("ActionLog.deleteById.start for category id = {}", id);
-        // check product count
-        cityRepository.deleteById(id);
-        log.info("ActionLog.deleteById.end for category id = {}", id);
+        CityEntity entity = cityRepository.findById(id).orElseThrow(() ->
+                new NotFoundException("CITY_NOT_FOUND"));
+        if(entity.getProductCount() > 0 )
+            throw new DeletionException("CITY_CANNOT_BE_DELETED");
+        else
+            cityRepository.deleteById(id);
     }
 }
